gen_rootfs: fix the wrong behavior of proprietary downloading

+++ gen_rootfs.sh
@@ -4,10 +4,16 @@ TOP=${PWD}
 
 # generate minimum rootfs
 gen_pure_rootfs() {
-  wget -c -t 0 --timeout=60 --waitretry=60 https://github.com/technexion-android/android_restricted_extra/raw/master/imx8-focol-proprietary.tar.gz
-  tar zxvf imx8-focol-proprietary.tar.gz
-  rm ./imx8-focol-proprietary.tar.gz
+
+  if [ ! -d rootfs_overlay ] ; then
+    wget -c -t 0 --timeout=60 --waitretry=60 https://github.com/technexion-android/android_restricted_extra/raw/master/imx8-focol-proprietary.tar.gz
+    tar zxvf imx8-focol-proprietary.tar.gz
+    rm ./imx8-focol-proprietary.tar.gz
+  fi
+
   mkdir rootfs
+
+  echo "generate ubuntu rootfs... default version: focal LTS"
   sudo debootstrap --arch=arm64 --keyring=/usr/share/keyrings/ubuntu-archive-keyring.gpg --verbose --foreign focal ${TOP}/rootfs
   sudo cp /usr/bin/qemu-aarch64-static ${TOP}/rootfs/usr/bin
   sudo LANG=C chroot ${TOP}/rootfs /debootstrap/debootstrap --second-stage
@@ -32,6 +38,8 @@ gen_pure_rootfs() {
   sudo mkdir -p ${TOP}/rootfs/etc/xdg/weston/
   sudo cp -a ${TOP}/rootfs_overlay/etc/xdg/weston/* ${TOP}/rootfs/etc/xdg/weston/
 
+  sudo cp -rv ${TOP}/rootfs_overlay/lib/firmware/* ${TOP}/rootfs/lib/firmware/
+
   sync
 
   sudo LANG=C chroot ${TOP}/rootfs /bin/bash -c "sudo ldconfig"
@@ -41,20 +49,6 @@ gen_pure_rootfs() {
   cd ${TOP}/rootfs
   sudo tar --exclude='./dev/*' --exclude='./lost+found' --exclude='./mnt/*' --exclude='./media/*' --exclude='./proc/*' --exclude='./run/*' --exclude='./sys/*' --exclude='./tmp/*' --numeric-owner -czpvf ../rootfs.tgz .
   cd ${TOP}
-}
-
-merge_tn_fs_overlay() {
-  wget -c -t 0 --timeout=60 --waitretry=60 https://github.com/technexion-android/android_restricted_extra/raw/master/imx8-focol-proprietary.tar.gz
-  tar zxvf imx6_7-bionic.tar.gz
-  mv ${TOP}/fs_overlay ${TOP}/
-  rm -rf ${TOP}/imx6_7-bionic.tar.gz
-  rm -rf ${TOP}/fs_overlay
-}
-
-throw_rootfs() {
-  sudo swapoff ${TOP}/rootfs/swapfile
-  sudo rm -rf ${TOP}/rootfs/
-  sudo rm rootfs.tgz
 }
 
 gen_pure_rootfs
